Move private traits to private.rs
